move common json-rpc bits to a library
