+++ .gitignore
@@ -88,3 +88,6 @@ data/meterpreter/ext_server_pivot.*.dll
 
 # local docker compose overrides
 docker-compose.local*
+
+# Ignore python bytecode
+*.pyc
